Add a storage accounts cache to Bank
