+++ programs/storage_api/Cargo.toml
@@ -17,9 +17,6 @@ serde_derive = "1.0.92"
 solana-logger = { path = "../../logger", version = "0.16.0" }
 solana-sdk = { path = "../../sdk", version = "0.16.0" }
 
-[dev-dependencies]
-solana-runtime = { path = "../../runtime", version = "0.16.0" }
-
 [lib]
 crate-type = ["lib"]
 name = "solana_storage_api"
