+++ programs/storage_program/Cargo.toml
@@ -14,6 +14,11 @@ solana-logger = { path = "../../logger", version = "0.16.0" }
 solana-sdk = { path = "../../sdk", version = "0.16.0" }
 solana-storage-api = { path = "../storage_api", version = "0.16.0" }
 
+[dev-dependencies]
+solana-runtime = { path = "../../runtime", version = "0.16.0" }
+assert_matches = "1.3.0"
+bincode = "1.1.4"
+
 [lib]
 crate-type = ["lib", "cdylib"]
 name = "solana_storage_program"
