+++ runtime/Cargo.toml
@@ -29,6 +29,7 @@ solana-noop-program = { path = "../programs/noop_program", version = "0.16.0" }
 solana-sdk = { path = "../sdk", version = "0.16.0" }
 solana-stake-api = { path = "../programs/stake_api", version = "0.16.0" }
 solana-stake-program = { path = "../programs/stake_program", version = "0.16.0" }
+solana-storage-api = { path = "../programs/storage_api", version = "0.16.0" }
 solana-vote-api = { path = "../programs/vote_api", version = "0.16.0" }
 solana-vote-program = { path = "../programs/vote_program", version = "0.16.0" }
 sys-info = "0.5.7"
